+++ example/windows/flutter/generated_plugins.cmake
@@ -3,7 +3,6 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-  screen_brightness_windows
   xue_hua_navite_video_player
 )