+++ lib/src/utils/temp_paths.dart
@@ -1,0 +1,1 @@
+export 'temp_paths_stub.dart' if (dart.library.io) 'temp_paths_io.dart';